refactor(core): merge init addon into core addon
