+++ addons/core/CfgFactionClasses.hpp
