+++ addons/core/CfgVehicles.hpp
