+++ addons/core/functions/fn_addPlayerActions.sqf
