+++ addons/core/functions/fn_addVehicleKeyWest.sqf
