+++ addons/core/functions/fn_initCommanderNPC.sqf
