+++ addons/core/functions/fn_initCoopRClients.sqf
