+++ addons/core/functions/fn_initCoopRServer.sqf
