+++ addons/core/functions/fn_initEventsClient.sqf
