+++ addons/core/functions/fn_initEventsServer.sqf
