+++ addons/core/functions/fn_initHQ.sqf
