+++ addons/core/functions/fn_initIntelligenceNPC.sqf
