+++ addons/core/functions/fn_initLobby.sqf
