+++ addons/core/functions/fn_initQuartermasterNPC.sqf
