+++ addons/core/functions/fn_initVehicleNPC.sqf
