+++ addons/core/functions/fn_setupBasicRoleLoadoutModule.sqf
