+++ addons/core/functions/fn_setupCommanderModule.sqf
