+++ addons/core/functions/fn_setupHQModule.sqf
