+++ addons/core/functions/fn_setupIntelligenceModule.sqf
