+++ addons/core/functions/fn_setupLobbyModule.sqf
