+++ addons/core/functions/fn_setupQuartermasterItemsModule.sqf
