+++ addons/core/functions/fn_setupQuartermasterModule.sqf
