+++ addons/core/functions/fn_setupVehicleModule.sqf
