+++ addons/core/functions/fn_verifySetup.sqf
