+++ addons/core/helper/fn_initPromise.sqf
