+++ addons/core/helper/fn_stringReplace.sqf
