+++ addons/core/helper/script_component.hpp
