+++ addons/core/script_component.hpp
@@ -4,4 +4,6 @@
 
 #define DEBUG_CTX "COOPR.CORE"
 
+#define SYNC_RATE_MIN 3
+
 #include "\x\coopr\addons\core\script_macros.hpp"
