+++ addons/core/stringtable.xml
