+++ addons/core/tests/fn_tests_init.sqf
