+++ addons/core/tests/script_component.hpp
