+++ addons/eventhandler/config.cpp
@@ -10,8 +10,7 @@ class CfgPatches {
 
         requiredVersion = 1.80;
         requiredAddons[] = {
-            "coopr_core",
-            "coopr_init"
+            "coopr_core"
         };
         // List of objects (CfgVehicles classes) contained in the addon. Important also for Zeus content unlocking.
         units[] = {};
