--- addons/init/script_component.hpp
@@ -1,9 +1,0 @@
-#include "\x\coopr\addons\core\script_mod.hpp"
-
-#define COMPONENT init
-
-#define DEBUG_CTX "COOPR.INIT"
-
-#define SYNC_RATE_MIN 3
-
-#include "\x\coopr\addons\core\script_macros.hpp"
