+++ addons/supply/helper/fn_hasEmptyCargo.sqf
