+++ build.cmd
@@ -15,7 +15,6 @@ mkdir %dest%
 %armake64% build -p addons\death %dest%\death.pbo
 %armake64% build -p addons\equipment %dest%\equipment.pbo
 %armake64% build -p addons\eventhandler %dest%\eventhandler.pbo
-%armake64% build -p addons\init %dest%\init.pbo
 %armake64% build -p addons\integration %dest%\integration.pbo
 %armake64% build -p addons\intel %dest%\intel.pbo
 %armake64% build -p addons\lobby %dest%\lobby.pbo
