+++ build.sh
@@ -14,7 +14,6 @@ $ARMAKE build -p -f addons/core $DEST/core.pbo
 $ARMAKE build -p -f addons/death $DEST/death.pbo
 $ARMAKE build -p -f addons/equipment $DEST/equipment.pbo
 $ARMAKE build -p -f addons/eventhandler $DEST/eventhandler.pbo
-$ARMAKE build -p -f addons/init $DEST/init.pbo
 $ARMAKE build -p -f addons/integration $DEST/integration.pbo
 $ARMAKE build -p -f addons/intel $DEST/intel.pbo
 $ARMAKE build -p -f addons/lobby $DEST/lobby.pbo
